+++ src/@types/global.d.ts
@@ -30,6 +30,7 @@ import {Notifier} from "../Notifier";
 import type {Renderer} from "react-dom";
 import RightPanelStore from "../stores/RightPanelStore";
 import WidgetStore from "../stores/WidgetStore";
+import CallHandler from "../CallHandler";
 
 declare global {
     interface Window {
@@ -53,6 +54,7 @@ declare global {
         mxNotifier: typeof Notifier;
         mxRightPanelStore: RightPanelStore;
         mxWidgetStore: WidgetStore;
+        mxCallHandler: CallHandler;
     }
 
     interface Document {
@@ -62,6 +64,7 @@ declare global {
 
     interface Navigator {
         userLanguage?: string;
+        mediaSession: any;
     }
 
     interface StorageEstimate {
